requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
stripe>=5.0
requests>=2.30
python-dotenv>=1.0
streamlit>=1.25